[FEATURE] Add URL crawler script
A new PHP script has been added to crawl a specified domain, extract all relevant URLs, and save them in a CSV file. This script filters out links to files, JavaScript, tel:, mailto:, and URLs with parameters. It can be used as an alternative to tools like the Screaming Frog SEO Spider.

The README.md file has been updated to include instructions on how to use this new crawler script. The create-backstop-scenarios.php file was also modified to read from the newly created CSV file by the crawler instead of the previous one.

Please note that the generated CSV file should be checked and possibly cleaned up manually before using it for tests.